Replace CacheableInteger with CacheableLong
git-svn-id: file:///home/svn/DVA2006@6097 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/commands/pilot/PilotCenterCommand.java
@@ -207,7 +207,7 @@ public class PilotCenterCommand extends AbstractTestHistoryCommand {
 				
 				// Get promotion queue size
 				GetPilotRecognition rdao = new GetPilotRecognition(con);
-				ctx.setAttribute("promoQueueSize", Integer.valueOf(rdao.hasPromotionQueue(myEQType)), REQUEST);
+				ctx.setAttribute("promoQueueSize", Long.valueOf(rdao.hasPromotionQueue(myEQType)), REQUEST);
 				
 				// Get exam/transfer queue sizes
 				GetExam exdao = new GetExam(con);
